Move visitor to separate files
